requirements/base.txt
beautifulsoup4
braintree
celery
culqipy
dateutils
dj-database-url
django
django-addanother
-e git+https://github.com/ikcam/django-autocomplete-light.git#egg=django-autocomplete-light
django-boilerplate
django-bootstrap4
django-ckeditor
django-countries
django-cors-headers
django-dotenv
django-filter
django-querysetsequence
django-redis
djangorestframework
drf-nested-routers
facebook-sdk
Pillow
psycopg2-binary
requests
six
stripe
tldextract
